requirements.txt
Flask
geopy==2.4.0
Pillow==10.2.0
pyTelegramBotAPI==4.15.2
quart==0.19.4
aiohttp